Move ES-related code into its own module
