+++ corehq/apps/events/tests/test_es.py
@@ -1,0 +1,7 @@
+import doctest
+
+
+def test_doctests():
+    import corehq.apps.events.es as module
+    results = doctest.testmod(module, optionflags=doctest.ELLIPSIS)
+    assert results.failed == 0
